requirements.txt
pandas
scikit-learn
fairlearn
mlflow
pyarrow
matplotlib
numpy
pytest
pyyaml
